renamed files routes and specs
